rename assetList to assetListBody in molecules
